mait-mvp-glitch/backend/requirements.txt
fastapi
uvicorn
chromadb
pydantic
python-multipart
requests
google-genai
python-dotenv
# RAG dependencies
sentence-transformers
pypdf
python-docx
